pages and components routing (work in progress)
