+++ app/client/src/components/InitialAssessment.jsx
@@ -1,0 +1,10 @@
+const InitialAssessment = () => {
+  return (
+    <>
+      <h2>InitialAssessment</h2>
+      {/* <p>({user ? 'Child registered' : 'No child assessed yet'})</p> */}
+    </>
+  );
+};
+
+export default InitialAssessment;
