+++ app/client/src/components/Main.jsx
@@ -3,13 +3,15 @@ import { AuthData } from '../components/auth/AuthWrapper';
 
 const Main = () => {
   const { user } = AuthData();
+
   const logout = () => {
     window.open('http://localhost:3001/api/auth/logout', '_self');
   };
   return (
     <div>
-      <h1>Main</h1>
-      <h2>Hello {user.firstName}</h2>
+      <h2>Main</h2>
+      {/* <h3>Hello {user.firstName}</h3> */}
+      {/* <p>({!getUser ? 'Child registered' : 'No child assessed yet'})</p> */}
       <ul>
         <li>My own path</li>
         <li>Catalog</li>
