--- app/client/src/pages/InitialAssessment.jsx
@@ -1,5 +1,0 @@
-const InitialAssessment = () => {
-  return <div>InitialAssessment</div>;
-};
-
-export default InitialAssessment;
